added user to questionnaire services and controller

+++ Code/services/src/main/kotlin/pt/iselearning/services/service/questionnaires/QuestionnaireServices.kt
@@ -4,6 +4,7 @@ import org.modelmapper.ModelMapper
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.validation.annotation.Validated
+import pt.iselearning.services.domain.User
 import pt.iselearning.services.domain.questionnaires.Questionnaire
 import pt.iselearning.services.models.questionnaire.QuestionnaireChallengeModel
 import pt.iselearning.services.models.questionnaire.QuestionnaireModel
@@ -11,6 +12,7 @@ import pt.iselearning.services.models.questionnaire.output.QuestionnaireOutputMo
 import pt.iselearning.services.models.questionnaire.QuestionnaireWithChallengesModel
 import pt.iselearning.services.models.questionnaire.output.QuestionnaireChallengeOutputModel
 import pt.iselearning.services.repository.questionnaire.QuestionnaireRepository
+import pt.iselearning.services.util.checkIfLoggedUserIsResourceOwner
 import pt.iselearning.services.util.checkIfQuestionnaireExists
 import javax.validation.Valid
 import javax.validation.constraints.Positive
@@ -42,12 +44,14 @@ class QuestionnaireServices(
      * Create a questionnaire and adds it challenges.
      *
      * @param questionnaireWithChallengesModel object information
+     * @param loggedUser user that is calling the service
      * @return created questionnaire
      */
     @Validated
     @Transactional
-    fun createQuestionnaireWithChallenges(@Valid questionnaireWithChallengesModel: QuestionnaireWithChallengesModel): Questionnaire? {
+    fun createQuestionnaireWithChallenges(@Valid questionnaireWithChallengesModel: QuestionnaireWithChallengesModel, loggedUser: User): Questionnaire? {
         val questionnaire = convertToEntity(questionnaireWithChallengesModel.questionnaire)
+        questionnaire.creatorId = loggedUser.userId
 
         val createdQuestionnaire = questionnaireRepository.save(questionnaire)
         val questionnaireChallengeModel = QuestionnaireChallengeModel(
@@ -76,27 +80,46 @@ class QuestionnaireServices(
     /**
      * Get all user questionnaires.
      *
-     * @param userId unique identifier of user object
+     * @param loggedUser user that is calling the service
      * @return List of questionnaires objects
      */
     @Validated
-    fun getUserAllQuestionnaires(@Positive userId: Int) : List<Questionnaire> {
-        return questionnaireRepository.findAllByCreatorId(userId)
+    fun getUserAllQuestionnaires(@Valid loggedUser: User) : List<Questionnaire> {
+        return questionnaireRepository.findAllByCreatorId(loggedUser.userId!!)
+    }
+
+    /**
+     * Get questionnaire with its challenges by its unique identifier
+     *
+     * @param questionnaireId unique identifier of user object
+     * @return List of QuestionnaireOutputModel objects
+     */
+    @Validated
+    fun getQuestionnaireByIdWithChallenge(@Positive questionnaireId: Int): QuestionnaireOutputModel? {
+        val questionnaire = checkIfQuestionnaireExists(questionnaireRepository, questionnaireId)
+        val challenges = questionnaireChallengeServices.getAllQuestionnaireChallengeByQuestionnaireId(questionnaireId)
+
+        return QuestionnaireOutputModel(
+                questionnaireId,
+                questionnaire.description,
+                questionnaire.timer,
+                challenges.map { QuestionnaireChallengeOutputModel(it.challenge, it.languageFilter) }
+        )
     }
 
     /**
      * Update a questionnaire.
      *
      * @param questionnaireModel information to be updated
+     * @param loggedUser user that is calling the service
      * @return updated questionnaire
      */
     @Validated
-    fun updateQuestionnaireById(@Positive questionnaireId: Int, @Valid questionnaireModel: QuestionnaireModel): Questionnaire {
-        val questionnaireFromDB = questionnaireRepository.findById(questionnaireId)
-        checkIfQuestionnaireExists(questionnaireFromDB, questionnaireId)
+    fun updateQuestionnaireById(@Positive questionnaireId: Int, @Valid questionnaireModel: QuestionnaireModel, @Valid loggedUser: User): Questionnaire {
+        val updatedQuestionnaire = checkIfQuestionnaireExists(questionnaireRepository, questionnaireId)
+        checkIfLoggedUserIsResourceOwner(loggedUser.userId!!, updatedQuestionnaire.creatorId!!)
 
         //region data for update operation
-        val updatedQuestionnaire = questionnaireFromDB.get()
         if(questionnaireModel.description != null)
             updatedQuestionnaire.description = questionnaireModel.description
         if(questionnaireModel.timer != null)
@@ -110,23 +133,14 @@ class QuestionnaireServices(
      * Delete a questionnaire by its unique identifier.
      *
      * @param questionnaireId identifier of object
+     * @param loggedUser user that is calling the service
      */
     @Validated
-    fun deleteQuestionnaireById(@Positive questionnaireId: Int) {
-        checkIfQuestionnaireExists(questionnaireRepository, questionnaireId)
-        questionnaireRepository.deleteById(questionnaireId)
-    }
-
-    fun getQuestionnaireInstanceByIdWithChallenge(questionnaireId: Int): QuestionnaireOutputModel? {
+    fun deleteQuestionnaireById(@Positive questionnaireId: Int, @Valid loggedUser: User) {
         val questionnaire = checkIfQuestionnaireExists(questionnaireRepository, questionnaireId)
-        val challenges = questionnaireChallengeServices.getAllQuestionnaireChallengeByQuestionnaireId(questionnaireId)
+        checkIfLoggedUserIsResourceOwner(loggedUser.userId!!, questionnaire.creatorId!!)
 
-        return QuestionnaireOutputModel(
-                questionnaireId,
-                questionnaire.description,
-                questionnaire.timer,
-                challenges.map { QuestionnaireChallengeOutputModel(it.challenge, it.languageFilter) }
-        )
+        questionnaireRepository.deleteById(questionnaireId)
     }
 
     /**
